+++ CHANGELOG.md
@@ -1,13 +1,13 @@
 # Change Log
 
 ## Added
+- disable plugins setting
 - jwt leeway setting
 
 ## 7.6.7
 ## Added
 - Dutch translation
 - Chinese (Traditional, Taiwan), Basque (Spain), Armenian and Malay (Malaysia) empty file templates
-- disable plugins setting
 
 ## Changed
 - fix edit locked file