+++ CHANGELOG.md
@@ -8,6 +8,9 @@
 ## Changed
 - fix notification length
 
+## Changed
+- additional check availability for external files when mention
+
 ## 7.6.7
 ## Added
 - Dutch translation